+++ src/cert/x509_test.rs
@@ -6,15 +6,13 @@
 //!
 //! These are hung off to the side to avoid cluttering the main x509.rs.
 
-use ::ring::error::Unspecified;
-
 use crate::cert;
 use crate::cert::testdata;
 use crate::cert::Algo;
 use crate::cert::Cert;
 use crate::cert::CertFormat;
 use crate::cert::PublicKeyParams;
-use crate::crypto::ring;
+use crate::cert::RingCiphers;
 use crate::crypto::sig;
 
 /// A `Ciphers` that blindly accepts all signatures.
@@ -38,45 +36,6 @@ impl cert::Ciphers for NoVerify {
         _: &PublicKeyParams,
     ) -> Option<&'a mut dyn sig::Verify<Error = ()>> {
         Some(self)
-    }
-}
-
-/// A `Ciphers` built on top of `ring`.
-struct RingCiphers {
-    verifier: Option<Box<dyn sig::Verify<Error = Unspecified>>>,
-}
-impl RingCiphers {
-    fn new() -> Self {
-        Self { verifier: None }
-    }
-}
-impl cert::Ciphers for RingCiphers {
-    type Error = Unspecified;
-    fn verifier<'a>(
-        &'a mut self,
-        algo: Algo,
-        key: &PublicKeyParams,
-    ) -> Option<&'a mut dyn sig::Verify<Error = Unspecified>> {
-        match (key, algo) {
-            (
-                PublicKeyParams::Rsa { modulus, exponent },
-                Algo::RsaPkcs1Sha256,
-            ) => {
-                use crate::crypto::rsa::Builder as _;
-
-                let key = ring::rsa::PublicKey::new(
-                    (*modulus).into(),
-                    (*exponent).into(),
-                )
-                .unwrap();
-                let rsa = ring::rsa::Builder::new();
-                self.verifier = Some(Box::new(rsa.new_verifier(key).unwrap()));
-                match &mut self.verifier {
-                    Some(x) => Some(&mut **x),
-                    None => None,
-                }
-            }
-        }
     }
 }
